backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0
bcrypt==4.1.2
PyJWT==2.8.0
razorpay==1.4.1
python-dotenv==1.0.0
gunicorn==21.2.0
bleach==6.1.0
sendgrid==6.11.0
reportlab==4.0.7
